+++ components/studentPortal/freemium/modals/Upgrade.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+export default function Upgrade() {
+  return (
+    <div className="md:space-y-10 space-y-2">
+      Welcome to Skillify Coding Academy!
+    </div>
+  );
+}
