+++ components/studentPortal/freemium/modals/Welcome.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+export default function Welcome() {
+  return (
+    <div className="md:space-y-10 space-y-2">
+      Welcome to Skillify Coding Academy!
+    </div>
+  );
+}
