+++ components/studentPortal/freemium/modals/WhereToStart.tsx
@@ -1,0 +1,4 @@
+import React from "react";
+export default function WhereToStart() {
+  return <div className="md:space-y-10 space-y-2">Start Here.</div>;
+}
